+++ src/main/java/kitchenpos/dto/vo/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.vo;
+package kitchenpos.dto.vo;
 
 import java.math.BigDecimal;
 import java.util.Objects;
