+++ src/test/java/kitchenpos/fixture/ProductFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 @SuppressWarnings("NonAsciiCharacters")
 public class ProductFixture {
