+++ src/test/java/kitchenpos/repository/ProductRepositoryTest.java
@@ -3,7 +3,8 @@ package kitchenpos.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.repository.ProductRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
